StadiaSync/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
firebase-admin==6.4.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.1
stripe==8.5.0
pydantic==2.6.1
supabase==2.3.4
python-dotenv==1.0.1
